server/requirements.txt
Flask
gunicorn
minio
python-docx
requests